Update README.md
Bradley shop is a family-owned store that sells beverages.  The store offers 3 types of beverages: Coffee, Alcohol, and Smoothie. The store is open from 8 in the morning to 11 pm in the afternoon.  The owner of the shop likes to automate the order transactions and reports, and purchases software for testing order activities for one month. You are asked to implement this software based on the following requirements.